refactor: Simplify services list using flex-wrap
In this commit, the services list in the Services section of 3-index.html is simplified by removing the second ul and placing the three li elements directly under the first ul. The CSS file (3-styles.css) is updated with a new comment section for the SERVICES and a selector targeting the row class inside the section-services class is added with the property flex-wrap set to wrap.

Repo: alx_html_css
Directory: flexbox
Files: 3-index.html, 3-styles.css