feat: Align center logo and navbar
In this task, I implemented the alignment of the logo and navbar to be centered within the header container. The changes were made in the 7-styles.css file.

Inside the /* Header section */, I added the CSS property 'align-items: center' to the .header-container class selector. This change ensures that the logo and navbar items are vertically centered within the header container.

With this alignment update, the header now has a more visually appealing and balanced layout, enhancing the overall user experience.

GitHub repository: alx_html_css
Directory: flexbox
Files modified: 7-index.html, 7-styles.css